+++ src/main/java/seong/onlinestudy/controller/CommentController.java
@@ -2,7 +2,7 @@ package seong.onlinestudy.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import seong.onlinestudy.SessionConst;
+import seong.onlinestudy.controller.response.Result;
 import seong.onlinestudy.domain.Member;
 import seong.onlinestudy.exception.InvalidSessionException;
 import seong.onlinestudy.request.CommentCreateRequest;
