+++ src/main/java/seong/onlinestudy/controller/PostController.java
@@ -3,6 +3,7 @@ package seong.onlinestudy.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
+import seong.onlinestudy.controller.response.Result;
 import seong.onlinestudy.domain.Member;
 import seong.onlinestudy.domain.PostCategory;
 import seong.onlinestudy.dto.PostDto;
