+++ src/main/java/seong/onlinestudy/controller/RecordController.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.SessionAttribute;
+import seong.onlinestudy.controller.response.Result;
 import seong.onlinestudy.domain.Member;
 import seong.onlinestudy.dto.StudyRecordDto;
 import seong.onlinestudy.request.RecordsGetRequest;
@@ -25,7 +26,7 @@ public class RecordController {
 
     @GetMapping("/records")
     public Result<List<StudyRecordDto>> getRecords(@Valid RecordsGetRequest request,
-                             @SessionAttribute(value = LOGIN_MEMBER, required = false) Member loginMember) {
+                                                   @SessionAttribute(value = LOGIN_MEMBER, required = false) Member loginMember) {
         List<StudyRecordDto> records = recordService.getRecords(request, loginMember);
 
         return new Result<>("200", records);
