+++ src/main/java/seong/onlinestudy/controller/StudyController.java
@@ -3,6 +3,7 @@ package seong.onlinestudy.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
+import seong.onlinestudy.controller.response.Result;
 import seong.onlinestudy.domain.Member;
 import seong.onlinestudy.dto.StudyDto;
 import seong.onlinestudy.request.StudyCreateRequest;
