+++ src/main/java/seong/onlinestudy/controller/TicketMessageController.java
@@ -7,6 +7,7 @@ import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import seong.onlinestudy.controller.response.Result;
 import seong.onlinestudy.domain.Ticket;
 import seong.onlinestudy.dto.MemberTicketDto;
 import seong.onlinestudy.dto.TicketDto;
